+++ git-odb/src/store/linked/store/handle.rs
@@ -1,0 +1,33 @@
+use crate::linked::store::Handle;
+use git_features::threading::OwnShared;
+
+pub type PackCache = dyn git_pack::cache::DecodeEntry + Send + 'static;
+pub type NewPackCacheFn = dyn Fn() -> Box<PackCache> + 'static;
+
+pub type ObjectCache = dyn git_pack::cache::Object + Send + 'static;
+pub type NewObjectCacheFn = dyn Fn() -> Box<ObjectCache> + 'static;
+
+impl<S> Handle<S> {
+    pub fn with_pack_cache(mut self, create: impl Fn() -> Box<PackCache> + 'static) -> Self {
+        self.new_pack_cache = OwnShared::new(create);
+        self.pack_cache = (self.new_pack_cache)();
+        self
+    }
+    pub fn with_object_cache(mut self, create: impl Fn() -> Box<ObjectCache> + 'static) -> Self {
+        self.new_object_cache = OwnShared::new(create);
+        self.object_cache = (self.new_object_cache)();
+        self
+    }
+}
+
+impl<S: Clone> Clone for Handle<S> {
+    fn clone(&self) -> Self {
+        Handle {
+            store: self.store.clone(),
+            new_pack_cache: self.new_pack_cache.clone(),
+            new_object_cache: self.new_object_cache.clone(),
+            pack_cache: (self.new_pack_cache)(),
+            object_cache: (self.new_object_cache)(),
+        }
+    }
+}
